[Refactor] Move and restructure Gamemodes components
